requirements.txt
cryptowatch-sdk
boto3